Move auth.js file under auth folder
